telegram-mcp/pyproject.toml
[project]
name = "telegram-mcp"
version = "0.1.0"
description = "MCP server for Telegram channel monitoring with engagement scoring"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "telethon>=1.34.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
telegram-mcp = "telegram_mcp.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
